+++ user_docs/en/changes.md
@@ -26,6 +26,7 @@
 
 ### Bug Fixes
 
+* Resetting NVDA's configuration to factory defaults now also resets Add-on Store's settings. (#19866, @amirmahdifard)
 * The actions button can now be used when selecting multiple add-ons in the Add-on Store to perform batch actions, instead of just via the context menu in the add-ons list. (#19971, @amirmahdifard)
 * When moving to an ARIA grid cell in focus mode in web browsers, NVDA no longer reports both the row and column headers even if only the row or only the column changed. (#17750, @jcsteh)
 * In live text regions, such as terminals, NVDA no longer freezes when substantial amounts of text are dumped to the screen. (#20177)